+++ src/main/java/cua/integra/cajasuamc/controller/AlumnoController.java
@@ -1,0 +1,4 @@
+package cua.integra.cajasuamc.controller;
+
+public class AlumnoController {
+}
