+++ src/main/java/cua/integra/cajasuamc/entity/Alumno.java
@@ -1,0 +1,31 @@
+package cua.integra.cajasuamc.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Entity
+@Table(name="alumno")
+@Setter
+@Getter
+@NoArgsConstructor
+@ToString
+public class Alumno {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @Column(nullable = false, unique = true)
+    private int matricula;
+
+    @Column(nullable = false)
+    private String nombre;
+
+    @Column(nullable = false)
+    private String apellidos;
+
+    @Column(nullable = false)
+    private String telefono;
+}
