+++ src/main/java/cua/integra/cajasuamc/entity/Caja.java
@@ -1,0 +1,21 @@
+package cua.integra.cajasuamc.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+
+@Entity
+@Table(name="caja")
+@Setter
+@Getter
+@NoArgsConstructor
+@ToString
+public class Caja {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    @Column(nullable = false, unique = true)
+    private String numero_ventanilla;
+}
